requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
cloudscraper==1.2.71
fake-useragent==1.4.0